+++ src/TrashLib/Sonarr/IReleaseProfileLister.cs
@@ -1,0 +1,6 @@
+namespace TrashLib.Sonarr;
+
+public interface IReleaseProfileLister
+{
+    void ListReleaseProfiles();
+}